working mirroring meta-tiles

+++ dos/build-and-run.sh
@@ -1,6 +1,9 @@
 #!/bin/bash
+ecjo ""
+echo ">>> P1X ASM BUILD SCRIPT <<<"
+echo "============================"
+echo ""
 
-# Check if an argument is provided
 if [ $# -eq 0 ]; then
     echo "Usage: $0 <filename_without_extension> [boot]"
     exit 1
@@ -8,12 +11,20 @@ fi
 
 filename=$1
 
-fasm "${filename}/${filename}.asm" "game.com"
 
 if [ $# -eq 2 ]; then
     dd if=/dev/zero of=floppy.img bs=1474560 count=1
-    fasm boot.asm boot.bin
-    fasm "${filename}/${filename}.asm" "game.bin"
+    
+    if ! fasm boot.asm boot.bin; then
+        echo "Failed to assemble boot.asm"
+        exit 1
+    fi
+    
+    if ! fasm "${filename}/${filename}.asm" "game.bin"; then
+        echo "Failed to assemble ${filename}/${filename}.asm"
+        exit 1
+    fi
+
     dd if=boot.bin of=floppy.img bs=512 count=1 conv=notrunc
     dd if=game.bin of=floppy.img bs=512 seek=1 conv=notrunc
     qemu-system-i386 \
@@ -21,8 +32,15 @@ if [ $# -eq 2 ]; then
     -k en-us \
     -rtc base=localtime \
     -device cirrus-vga \
+    -cpu 486 \
+    -boot a \
     -fda floppy.img
 else
+    if ! fasm "${filename}/${filename}.asm" "game.com"; then
+        echo "Failed to assemble ${filename}/${filename}.asm"
+        exit 1
+    fi
+    
     qemu-system-i386 \
     -m 16 \
     -k en-us \
@@ -30,5 +48,6 @@ else
     -device cirrus-vga \
     -fda freedos.img \
     -drive file=fat:rw:. \
+    -cpu 486 \
     -boot order=a
 fi
